asiotest: separeted counters to tools_helper.hpp
